Fix issue where save was impossible after copy previous
Apparently, it really is the case that Hibernate uses

1. an update: set SpecimenId = null
2. a delete query

to delete a relation.
That's why the unique key was unfavored.
Unfortunately, this brings the todo, that the LatLong has to be
cleaned up when improving the database structure, as currently,
in the database/Hibernate config, it is handled as many-to-one,
in the code it is a one-to-one, imposing possible problems in the future.